Добавлены явные преобразования типов

+++ libs/utils/uniio.c
@@ -165,9 +165,9 @@ static int out_func_buffer(universal_io *const io, const char *const format, va_
 
 	int ret = vsnprintf(&io->out_buffer[io->out_position], io->out_size - io->out_position, format, local);
 
-	if (ret != -1 && ret + io->out_position < io->out_size)
+	if (ret != -1 && (size_t)ret + io->out_position < io->out_size)
 	{
-		io->out_position += ret;
+		io->out_position += (size_t)ret;
 		return ret;
 	}
